Refactor: extract CurveTail base class
